fix: reject fractional class_mapping values, pin two more live behaviors
class_mapping used z.number() instead of z.number().int(), so a
fractional value would have passed local validation. Added a live-
captured second stop action fixture (none, from a start/stop race)
and a live test attempt for a dataset with no classes, which found
the platform actually refuses with "Dataset must be ready to
auto-annotate" rather than the spec's documented 422.

+++ src/tools/index.ts
@@ -532,7 +532,7 @@ export const TOOL_DEFINITIONS: readonly ToolDefinition[] = [
           "IoU threshold for generated labels. Omit to use the server default (currently 0.7).",
         ),
       class_mapping: z
-        .array(z.union([z.number(), z.null()]))
+        .array(z.union([z.number().int(), z.null()]))
         .optional()
         .describe(
           "Model class index -> dataset class index mapping, positioned by model class index. Required to bridge a class-taxonomy mismatch between the model and the dataset; passed through with no length check.",
